Implements log in screen and oauth/token methods for getting an access token from user/pass or a refresh token. Fixes swapped time/distance filter on location update request

+++ src/com/geoloqi/android1/LQLocationData.java
@@ -58,7 +58,7 @@ public class LQLocationData extends SQLiteOpenHelper {
 		}
 	}
 
-	public void addLocation(Location location, int distanceFilter, int trackingLimit, int rateLimit) {
+	public void addLocation(Location location, float distanceFilter, long trackingLimit, int rateLimit) {
 		SQLiteDatabase db = this.getWritableDatabase();
 		ContentValues values = new ContentValues();
 		values.put(DATE, (int)(System.currentTimeMillis() / 1000));
